midas/requirements.txt
aiohttp==3.6.1
async-timeout==3.0.1
attrs==19.2.0
certifi==2019.9.11
chardet==3.0.4
cycler==0.10.0
Django==2.2.6
django-slack-oauth==1.5.0
idna==2.8
jsonfield==2.0.2
kiwisolver==1.1.0
matplotlib==3.1.1
multidict==4.5.2
numpy==1.17.2
opencv-python==4.1.1.26
Pillow==6.2.0
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.3
requests==2.22.0
six==1.12.0
slackclient==2.2.1
sqlparse==0.3.0
urllib3==1.25.6
yarl==1.3.0
